feat(datenraum-integration): Add static renderer
