+++ packages/editor/src/plugins/datenraum-integration/state.ts
@@ -1,0 +1,13 @@
+import { object, optional, string } from '../../plugin'
+
+export const state = object({
+  resource: optional(
+    object({
+      url: string(),
+      title: string(),
+      description: string(),
+    })
+  ),
+})
+
+export type DatenraumIntegrationState = typeof state
